Update EventsList menu and layout.
Move events list menu outside of events list, as it is used to control the events list.
Change behavior of PinMap clicks so that it highlights events in the list of that location, rather than jumping to that event.
Update layout of EventsList so that PinMap is side by side on larger screens.