Support Clubs. Support bulk event creation UI. Refactor and reorganize forms and pages.
Add support for creating and selecting clubs (for the creation of bulk events).
Create location and create event pages/forms are reorganized. The forms are now a re-usable organism component. SelectLocation component is added as the entry point for CreateEvent and CreateClubEvents. CreateEvent in the future will also be renamed CreateClubEvent.