Improve AddLeagueMatch tool.
Fix handling of new players for ManagedSession so that unrated players are initialized with default ratings.
Refactor handling of selected players so that it uses the Graphql RSVP node data for the player, composed into team  and match types instead of the previous winning/losing players.
Add support for a queue of matches that can be directly selected manually or from the matchmaking component.