Fix bugs with ratings calculations.
Fix behavior of completing a match so that the winner/loser is properly decided. Refactor the logic to go through the onSubmit callback. The bug was causing completed games to encode the wrong winner.
Fix bug with manual team widget. The widget player data was not being updated because of local state.
Fix rendering of RsvpPlayer skill bar due to divide by zero error. It will preserve the bar when the min and max skill value is 0.