Upgrade UI with Tailwind Catalyst.
Use Catalyst stacked layout.
Improve large screen split layout with side by side map and list of events.
Private events are hidden by default, but with inline expanding possible.
Move RSVP status and players list to the top of the page. Hide the players list by default.
Refactor routes to be grouped together logically.
Fix pagination of Club, Location, and ViewerEvents pages.
Fix afterDate param when querying for events so that if the value is not specified it will query for after the current date on the server side. The client-side generation of the current time causes rendering mismatch, so the null/None value indicates "after now".