Add Round Robin matchmaking.
Finish implementation of round-robin generator. It generates matches and also tries to uniformly shuffle the players (implementation is still slightly flawed). Round Robin match scheduling can be done without consuming the players so all the round-robin matches can be added to the schedule.
Fix Drag and Drop in match view when duplicate players exist (in the case of scheduling multiple matches for the same players).