Add support for minimum event ratings.
Fix edit event form max rsvps input so it can be empty on initial load. The number inputs would not accept an empty value when the field is not dirty (modified).
Add support for minimum rating for events.
Upgrade ppx/rhf.
Switch implementation of the AiTetsu player queue to use an ordered list rather than an unordered set.