Competitive Plus match making always prioritizes the top player.
Previously comp plus may suggest the other matches if matches with the top player is not optimal. This would force the top player to always be put into a suboptimal match with remaining players. Now the top player in the queue is always at the top of the suggested matches, regardless of match quality.