Improve FairPlay.
Support tap to select winner with optional score input. Un-scored results are encoded as 1 and -1 for the score input.
Add support for tap to replace a player in the Matches view.
Add support for requiring a set of players in draws in CompMatch.
Minor UI improvements such as in the tabs and convert some UiAction components to Buttons. Remove some unused old UI in the settings screen.