Add support for Anti-Teams.
Add support for Anti-Teams that uses the old "avoid all players" logic. Anti-teams are groups of players that should never play together in the same match.
Improve the UI of the selected players drawer.
Fix ActionBar on small screens. The number of courts input is hidden if a player selection is active.