Fix fixed-pair matchmaking.
Fixed-pair matchmaking was failing for Competitive Plus strategy if the fixed pair skill levels transcend the k-means group. Also the concept of a fixed pair is at odds with competitive matchmaking because a fixed-pair can be rigged to be "non-competitive." Now fixed-pairs are always optimized using the Mixed matchmaking strategy. Once the fixed pair matches are made it resumes with the selected strategy.
Preliminary support for new club events page and club events share link.