Add play count display and fix play count bug.
Fix bug with play count calculation. The first round of play will cause the average play count to be < 1. The next round of play will "boost" the 0 play count people in the round to a 0 play count. We now set the play count as minimum of 1 to avoid this issue. After the average play count is > 1, anyone with 0 play count joining will get boosted from that point.
Add the play count of each player to the Queue screen.