Add support for moving user off pending list. Improve score entry UI.
Refactor the EventRsvp component to have the RsvpOptions module rather than the EventRsvpUser component. This gives it access to the underlying Rsvp rather than just the User, which allows the update listType mutation to work.
Score entry in FairPlay now uses a number input on mobile so full keyboard is avoided.