Fix FairPlay bugs.
Fix TinyBase player data desync. Player data in TinyBase is not updated after a match is completed. Instead it is synced when a match is queued, but the logic had a bug where it only updated the player if the player record didn't exist yet. It now syncs the player data every time a match is queued.
Fix bug with progression of play count when not in session mode. Play count was being incremented extra times due to an extra Array.map loop.
Remove Storybook stories for Event and EventRsvps.