Add AI LLM Chat for event creation.
New Add an Event button launches the AI assistant for creating events. The assistant can work with natural language to suggest 1 or multiple events to be created. The button can be provided context from the page it is on. On the home page it creates events for the selected activity. On club pages it will use the default activity and club for events.
Refactor routes to force selection of an activity. The default route now shows pickleball rather than all events for all activities.